% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempind.R
\name{TemporalIndicators}
\alias{TemporalIndicators}
\title{Create temporal indicators}
\usage{
TemporalIndicators(
  Dates,
  DATEnum = TRUE,
  DOW = TRUE,
  week_start_in = 1,
  YEAR = TRUE,
  DOY = TRUE,
  WEEK = TRUE,
  NWD = "Individual",
  SWAP = TemporalIndicators::SwappedWorkingDays,
  Congresses = NULL,
  CongressDatesIn = TemporalIndicators::CongressDates,
  CongPrePost = TRUE,
  SDAY = TRUE,
  MONTH = TRUE,
  SEASON = TRUE
)
}
\arguments{
\item{Dates}{The vector of dates from which the indicators should be calculated}

\item{DATEnum}{Numeric dates needed?}

\item{DOW}{Day-of-week indicator needed?}

\item{week_start_in}{Day on which week starts (1: Monday, 7: Sunday)}

\item{YEAR}{Year indicator needed?}

\item{DOY}{Day-of-year indicator needed?}

\item{WEEK}{Week indicator needed?}

\item{NWD}{Non-working day indicator needed? (NULL if not, "None" if needed but without pre/post indicators, "Global"
if needed with one single pre/post indicator, regardless of the non-working day, "Individual" if an own pre/post indicator
is needed for each non-working day individually)}

\item{SWAP}{Swapped working day indicator needed? (NULL if not, a data frame with columns 'date' (date of swap) and 'name'
(names, e.g. 'Non-working day' and 'Working day' ) if yes)}

\item{Congresses}{What congress indicators are needed? (NULL if none, list of congresses otherwise)}

\item{CongressDatesIn}{Data frame that holds the dates of congresses with columns 'CongressName' (name of the congress),
'StartDate' (starting date), 'EndDate' (end date)}

\item{CongPrePost}{Pre/post indicators for the congress(es) needed?}

\item{SDAY}{Semmelweis Day indicator needed?}

\item{MONTH}{Month indicator needed?}

\item{SEASON}{Season indicator needed?}
}
\value{
A data frame containing the requested indicators in its columns
}
\description{
Create temporal indicators
}
\examples{
TemporalIndicators( seq( as.Date( "2018-01-01" ), as.Date( "2019-03-31" ), by = "days" ) )
}
